gcp/inference/requirements.txt
google-cloud-storage
netCDF4
pandas
rasterio
xarray